removed some warnings

+++ Cheat Engine/ProcessWindowUnit.pas
@@ -557,7 +557,7 @@ begin
   begin
     DBKFileAsMemory(opendialog2.filename);
     processselected:=true;
-    ProcessHandler.ProcessHandle:=-1;
+    ProcessHandler.ProcessHandle:=QWORD(-1);
     MainForm.ProcessLabel.caption:=extractfilename(opendialog2.FileName);
     MainForm.miSaveFile.visible:=true;
     ProcessHandler.processid:=$FFFFFFFF;
